Isolate the dummy login to a seperate interface
